pyproject.toml
[project]
name = "epistemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
